+++ config/filesystems.php
@@ -46,7 +46,7 @@ return [
 
         'bar-assistant' => [
             'driver' => 'local',
-            'root' => storage_path('uploads'),
+            'root' => storage_path('bar-assistant/uploads'),
             'url' => env('APP_URL').'/uploads',
             'visibility' => 'public',
             'throw' => false,
@@ -79,7 +79,7 @@ return [
 
     'links' => [
         public_path('storage') => storage_path('app/public'),
-        public_path('uploads') => storage_path('uploads'),
+        public_path('uploads') => storage_path('bar-assistant/uploads'),
     ],
 
 ];