+++ resources/docker/run.sh
@@ -4,18 +4,14 @@ set -e
 
 first_time_check() {
     if [ ! -f /var/www/cocktails/.env ]; then
-        echo "Application .env file not found, creating a new .env file..."
-
-        cd /var/www/cocktails
-
         cp .env.dist .env
 
         php artisan key:generate
         php artisan storage:link
 
-        if [ ! -f /var/www/cocktails/storage/database.sqlite ]; then
+        if [ ! -f /var/www/cocktails/storage/bar-assistant/database.sqlite ]; then
             echo "Database not found, creating a new database..."
-            touch /var/www/cocktails/storage/database.sqlite
+            touch /var/www/cocktails/storage/bar-assistant/database.sqlite
             php artisan migrate --force
             php artisan bar:open
         else
@@ -26,6 +22,7 @@ first_time_check() {
 }
 
 start_system() {
+    mkdir -p /var/www/cocktails/storage/bar-assistant/uploads/{cocktails,ingredients,temp}
     first_time_check
 
     php artisan bar:refresh-search
@@ -35,9 +32,9 @@ start_system() {
     php artisan config:cache
     php artisan route:cache
 
-    echo "Application ready!"
+    echo "!***************************!"
+    echo "!    Application ready      !"
+    echo "!***************************!"
 }
 
 start_system
-
-exec apache2-foreground